requirements.txt
autopep8==2.3.1
black==24.10.0
click==8.1.8
coverage==7.6.10
flake8==7.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pygame==2.6.1
setuptools==75.6.0
setuptools-scm==8.1.0
websocket-client==1.8.0
websockets==14.1
